move AfterSchool related components to AfterSchool folder
